moved filling machine logic to rocketstate service
