+++ invictus2/obc/include/services/rocket_state.h
@@ -1,0 +1,9 @@
+#ifndef ROCKET_STATE_H_
+#define ROCKET_STATE_H_
+
+#include <stdbool.h>
+
+bool rocket_state_service_setup(void);
+void rocket_state_service_start(void);
+
+#endif // ROCKET_STATE_H_
