+++ invictus2/obc/src/services/rocket_state.c
@@ -1,0 +1,1 @@
+#include "services/rocket_state.h"
